# server/docs/swagger.yaml
definitions:
  request.ReqArr:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /user-basic/add:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      - description: 用户密码
        in: query
        name: password
        required: true
        type: string
      - description: 确认密码
        in: query
        name: rePassword
        required: true
        type: string
      - description: 手机号
        in: query
        name: phone
        required: true
        type: string
      - description: 邮箱
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户基础信息-用户添加
      tags:
      - user-basic(用户信息)
  /user-basic/delete:
    delete:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 用户信息参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqArr'
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户基础信息-用户删除
      tags:
      - user-basic(用户信息)
  /user-basic/query:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 用户信息参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqArr'
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户基础信息-用户查询
      tags:
      - user-basic(用户信息)
  /user-basic/update:
    put:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 用户id
        in: query
        name: id
        required: true
        type: string
      - description: 用户名字
        in: query
        name: name
        required: true
        type: string
      - description: 用户电话
        in: query
        name: phone
        required: true
        type: string
      - description: 用户邮箱
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户基础信息-用户修改
      tags:
      - user-basic(用户信息)
swagger: "2.0"